Remove old `Select` component and switch `SelectNew` component name
