+++ app/src/lib/settings/PreferencesForm.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
+	import Select from '$lib/Select/Select.svelte';
 	import SelectItem from '$lib/Select/SelectItem.svelte';
-	import SelectNew from '$lib/Select/SelectNew.svelte';
 	import { GitConfigService } from '$lib/backend/gitConfigService';
 	import { Project, ProjectService } from '$lib/backend/projects';
 	import SectionCard from '$lib/components/SectionCard.svelte';
@@ -151,7 +151,7 @@
 	</SectionCard>
 	{#if signCommits}
 		<SectionCard orientation="column">
-			<SelectNew
+			<Select
 				value={signingFormat}
 				options={signingFormatOptions}
 				label="Signing format"
@@ -166,7 +166,7 @@
 						{item.label}
 					</SelectItem>
 				{/snippet}
-			</SelectNew>
+			</Select>
 
 			<TextBox
 				label="Signing key"
